+++ src/main/java/databute/databuter/cluster/handshake/request/HandshakeRequestMessageSerializer.java
@@ -1,4 +1,4 @@
-package databute.databuter.cluster.handshake;
+package databute.databuter.cluster.handshake.request;
 
 import databute.databuter.network.message.MessageSerializer;
 import databute.databuter.network.packet.BufferedPacket;
