+++ src/main/java/databute/databuter/cluster/handshake/response/HandshakeResponseMessage.java
@@ -1,33 +1,20 @@
-package databute.databuter.cluster.handshake;
+package databute.databuter.cluster.handshake.response;
 
 import com.google.common.base.MoreObjects;
 import databute.databuter.cluster.network.ClusterMessageCode;
 import databute.databuter.network.message.Message;
 import databute.databuter.network.message.MessageCode;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
-public class HandshakeRequestMessage implements Message {
-
-    private final String id;
-
-    public HandshakeRequestMessage(String id) {
-        this.id = checkNotNull(id, "id");
-    }
+public class HandshakeResponseMessage implements Message {
 
     @Override
     public MessageCode messageCode() {
         return ClusterMessageCode.HANDSHAKE;
     }
 
-    public String id() {
-        return id;
-    }
-
     @Override
     public String toString() {
         return MoreObjects.toStringHelper(this)
-                .add("id", id)
                 .toString();
     }
 }
